feat(inventory): add inventory management API with validation, resources, and code cleanup (#10)
* feat(inventory): implement inventory management features including import, export, adjustment, and transaction history

* feat(inventory): add soft delete functionality to products, product units, and users

+++ app/Http/Resources/Products/ProductResource.php
@@ -23,7 +23,7 @@ class ProductResource extends JsonResource
             'description' => $this->description,
             'price' => (float)$this->price,
             'cost' => (float)$this->cost,
-            'stock' => (int)$this->stock,
+            'stock' => (float)$this->stock,
             'min_stock' => (int)$this->min_stock,
             'expiry_date' => $this->expiry_date?->format('Y-m-d'),
             'image' => $this->image ? asset($this->image) : null,
